work login and signup,  edit names of activities and fix visual bugs
