+++ app/src/main/java/com/icesi/umarket/SellerLoginActivity.kt
@@ -23,7 +23,7 @@ class SellerLoginActivity : AppCompatActivity() {
         binding.signUpHyperLink.setOnClickListener {
             var userType = intent.extras?.getString("userType","")
             if(userType=="consumer"){
-                startActivity(Intent(this,ConsumerRegistrationActivityData::class.java))
+                startActivity(Intent(this,ConsumerSignupActivity::class.java))
             }else{
                 startActivity(Intent(this,SellerSignupActivity::class.java))
             }
