+++ app/src/main/res/layout/activity_consumer_registration_photo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#5A51CE"
-    tools:context=".ConsumerRegistrationActivityPhoto">
+    tools:context=".AdditionalConsumerInfoActivity">
 
     <TextView
         android:id="@+id/textView2"
